requirements.txt
python-telegram-bot==20.0
tensorflow==2.10
scikit-learn==1.0.1